+++ packages/bpk-component-scrollable-calendar/src/BpkScrollableCalendarGridList-test.js
@@ -18,7 +18,7 @@
 
 import PropTypes from 'prop-types';
 import React from 'react';
-import isWeekend from 'date-fns/is_weekend';
+import isWeekend from 'date-fns/isWeekend';
 import { DateUtils } from 'bpk-component-calendar';
 import { colorPanjin } from 'bpk-tokens/tokens/base.es6';
 import { shallow } from 'enzyme';