+++ packages/bpk-tokens/formatters/bpk.default.scss.js
@@ -22,9 +22,16 @@ import sortTokens from './sort-tokens';
 import { blockComment } from './license-header';
 import { sassDocTemplate, nameTemplate, valueTemplate } from './bpk.scss';
 
-export const variableTemplate = ({ name, value, type }) =>
-  `${nameTemplate({ name })}: ${valueTemplate({ value, type })} !default;`;
+export const variableTemplate = ({ name, value, type }) => {
+  if (type === 'function') {
+    return `${value}`;
+  }
 
+  return `${nameTemplate({ name })}: ${valueTemplate({
+    value,
+    type,
+  })} !default;`;
+};
 export const template = ({ category, name, value, type }) =>
   `${sassDocTemplate({ category })}\n${variableTemplate({
     name,