+++ packages/bpk-tokens/formatters/bpk.es6.js.js
@@ -21,9 +21,12 @@ import _ from 'lodash';
 import sortTokens from './sort-tokens';
 import { blockComment } from './license-header';
 
-export const tokenTemplate = ({ name, value }) =>
-  `export const ${_.camelCase(name)} = "${value.replace(/"/g, '\\"')}";`;
-
+export const tokenTemplate = ({ name, value, type }) => {
+  if (type === 'function') {
+    return null;
+  }
+  return `export const ${_.camelCase(name)} = "${value.replace(/"/g, '\\"')}";`;
+};
 export const categoryTemplate = (
   categoryName,
   props,
@@ -39,7 +42,10 @@ export default result => {
     .uniq()
     .value();
 
-  const singleTokens = _.map(props, prop => tokenTemplate(prop)).join('\n');
+  const singleTokens = _.map(
+    props.filter(prop => prop.type !== 'function'),
+    prop => tokenTemplate(prop),
+  ).join('\n');
 
   const groupedTokens = categories
     .sort()
@@ -48,6 +54,7 @@ export default result => {
         category,
         _(props)
           .filter({ category })
+          .filter(prop => prop.type !== 'function')
           .value(),
       ),
     )